Backend/requirements.txt
# Core
pandas==2.1.4
numpy==1.26.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3

# Model Persistence
joblib==1.3.2

# Explainability
shap==0.44.1

# FastAPI
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# OCR & Text Processing
opencv-python==4.9.0.80

# LLM
groq==1.0.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Testing
pytest==8.1.1
